pyproject.toml
[project]
name = "trip-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "openai>=1.107.2",
    "pandas>=2.3.2",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "redis>=6.4.0",
    "uvicorn[standard]>=0.35.0",
]
